make the bitwidth the bit width
Previously, I was confusing BitWidth and Width. However, while it's technically
possible to allow arbitrary "widths", a AMT is defined as a prefix tree.
Therefore, it should be configured in terms of "bit" width (how many bits are
popped off the index at every level). This means valid _widths_ are confined to
powers of two.